requirements.txt
langchain==1.2.0
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
python-dotenv
chromadb
openai
